rename 'mags'
