+++ expo/src/MasloPersonaExpo.tsx
@@ -159,8 +159,8 @@ export class MasloPersonaExpo extends React.Component<Props, CompState> {
         reaction(_ => this.props.context.view, v => {
           this._updatePersonViewState(v);
         }),
-        reaction(_ => this.props.context.qolMags, qolMags => {
-          this._persona.updateDomainMags(qolMags);
+        reaction(_ => this.props.context.qolArmMagnitudes, qolArmMagnitudes => {
+          this._persona.updateDomainMagnitudes(qolArmMagnitudes);
         }),
       ];
       this._contextObserverDispose = () => {
