+++ lib/persona.armMagnitudes.ts
